bugfix for: Bug 511384 - in the oneM2M tests there is a situation where fields of the anytype type are not recognised in the Designer.
In some cases it was possible to reference the field of an anytype from a different module before the list of their fields was calculated. A good way to solve this problem seems to be to let the anytype calculate the list of its fields at the time of semantic checking. Making the semantic dependency explicit.

Signed-off-by: Kristof Szabados <Kristof.Szabados@ericsson.com>