adding support for calculating ans uding generated names (the names in the generated code are not always the same in the TTCN-3 code).
Please note set_genname was moved from the end of semantic checking to the beginning of the code generation, as I expect more semantic checking cycles than code generations to happen.
Also note that we do not need or should use the "const_", "modulepar_", "template_" prefixes ... bit I kept them as it is not yet completely clear if this would create problems in the future or not.

Signed-off-by: kristof <Kristof.Szabados@ericsson.com>